+++ fluxcloud/main/client.py
@@ -42,17 +42,12 @@ class ExperimentClient:
         if res.returncode != 0:
             raise ValueError("nonzero exit code, exiting.")
 
-    def run_command(self, cmd, cleanup_func=None):
+    def run_command(self, cmd):
         """
         Run a timed command, and handle nonzero exit codes.
         """
         logger.debug("\n> Running Command: " + " ".join(cmd))
         res = utils.run_command(cmd)
-
-        # An optional cleanup function (also can run if not successful)
-        if cleanup_func is not None:
-            cleanup_func()
-
         if res.returncode != 0:
             raise ValueError("nonzero exit code, exiting.")
 
@@ -69,17 +64,24 @@ class ExperimentClient:
         3. bring down the cluster
         """
         # Each experiment has its own cluster size and machine type
-        for experiment in setup.matrices:
-
-            # Don't bring up a cluster if experiments already run!
-            if not setup.force and experiment.is_run():
-                logger.info(
-                    f"Experiment on machine {experiment.expid} was already run and force is False, skipping."
-                )
-                continue
-
+        for experiment in setup.iter_experiments():
             self.up(setup, experiment=experiment)
             self.apply(setup, experiment=experiment)
+            self.down(setup, experiment=experiment)
+
+    @save_meta
+    def batch(self, setup):
+        """
+        Run Flux Operator experiments via batch submit
+
+        1. create the cluster
+        2. run each command via submit to same MiniCluster
+        3. bring down the cluster
+        """
+        # Each experiment has its own cluster size and machine type
+        for experiment in setup.iter_experiments():
+            self.up(setup, experiment=experiment)
+            self.submit(setup, experiment=experiment)
             self.down(setup, experiment=experiment)
 
     @save_meta
@@ -154,7 +156,7 @@ class ExperimentClient:
             # Ensure our credentials still work, and open port forward
             api.check(experiment)
 
-            # Get times back OR save as we go? (probably better)
+            # Save times (and logs in submit) as we go
             for jobid, info in api.submit(setup, experiment, size):
                 logger.info(f"{jobid} took {info['runtime']} seconds.")
                 self.times[jobid] = info["runtime"]
@@ -173,73 +175,57 @@ class ExperimentClient:
         This is really just running the setup!
         """
         # The MiniCluster can vary on size
-        minicluster = experiment.minicluster
         if not experiment.jobs:
             logger.warning(
                 f"Experiment {experiment.expid} has no jobs, nothing to run."
             )
             return
 
-        # The experiment is defined by the machine type and size
+        # Save output here
         experiment_dir = experiment.root_dir
 
-        # Iterate through all the cluster sizes
-        for size in minicluster["size"]:
+        for size, jobname, job in experiment.iter_jobs():
 
-            # We can't run if the minicluster > the experiment size
-            if size > experiment.size:
+            # Add the size
+            jobname = f"{jobname}-minicluster-size-{size}"
+            job_output = os.path.join(experiment_dir, jobname)
+            logfile = os.path.join(job_output, "log.out")
+
+            # Any custom commands to run first?
+            if hasattr(self, "pre_apply"):
+                self.pre_apply(experiment, jobname, job)
+
+            # Do we have output?
+            if os.path.exists(logfile) and not setup.force:
+                relpath = os.path.relpath(logfile, experiment_dir)
                 logger.warning(
-                    f"Cluster of size {experiment.size} cannot handle a MiniCluster of size {size}, skipping."
+                    f"{relpath} already exists and force is False, skipping."
                 )
                 continue
 
-            # Jobname is used for output
-            for jobname, job in experiment.jobs.items():
+            elif os.path.exists(logfile) and setup.force:
+                logger.warning(f"Cleaning up previous run in {job_output}.")
+                shutil.rmtree(job_output)
 
-                # Do we want to run this job for this size and machine?
-                if not experiment.check_job_run(job, size):
-                    logger.debug(
-                        f"Skipping job {jobname} as does not match inclusion criteria."
-                    )
-                    continue
+            # Create job directory anew
+            utils.mkdir_p(job_output)
 
-                # Add the size
-                jobname = f"{jobname}-minicluster-size-{size}"
-                job_output = os.path.join(experiment_dir, jobname)
-                logfile = os.path.join(job_output, "log.out")
+            # Generate the populated crd from the template
+            crd = experiment.generate_crd(job, size)
 
-                # Any custom commands to run first?
-                if hasattr(self, "pre_apply"):
-                    self.pre_apply(experiment, jobname, job)
+            # Prepare specific .crd for template
+            # Note the output directory is already specific to the job index
+            kwargs = {
+                "minicluster": experiment.minicluster,
+                "logfile": logfile,
+                "crd": crd,
+            }
+            apply_script = experiment.get_shared_script(
+                "minicluster-run", kwargs, suffix=f"-{jobname}"
+            )
 
-                # Do we have output?
-                if os.path.exists(logfile) and not setup.force:
-                    logger.warning(
-                        f"{logfile} already exists and force is False, skipping."
-                    )
-                    continue
-
-                elif os.path.exists(logfile) and setup.force:
-                    logger.warning(f"Cleaning up previous run in {job_output}.")
-                    shutil.rmtree(job_output)
-
-                # Create job directory anew
-                utils.mkdir_p(job_output)
-
-                # Generate the populated crd from the template
-                crd = experiment.generate_crd(job, size)
-
-                # Prepare specific .crd for template
-                # Note the output directory is already specific to the job index
-                kwargs = {"minicluster": minicluster, "logfile": logfile, "crd": crd}
-                apply_script = experiment.get_shared_script(
-                    "minicluster-run", kwargs, suffix=f"-{jobname}"
-                )
-
-                # Apply the job, and save to output directory
-                self.run_timed(
-                    f"{self.job_prefix}-{jobname}", ["/bin/bash", apply_script]
-                )
+            # Apply the job, and save to output directory
+            self.run_timed(f"{self.job_prefix}-{jobname}", ["/bin/bash", apply_script])
 
     def clear_minicluster_times(self):
         """